# openapi.json
{"openapi":"3.1.0","info":{"title":"Pooli API","description":"Pooli backend API docs","contact":{"name":"Pooli Team"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"User Policy","description":"User policy APIs"},{"name":"Admin Policy","description":"Admin policy APIs"}],"paths":{"/api/policies":{"get":{"tags":["User Policy"],"summary":"List active policies","description":"Returns active policies that a family representative can apply to the family group.","operationId":"getActivePolicies","responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivePolicyResponse"}}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivePolicyResponse"}}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivePolicyResponse"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivePolicyResponse"}}}}}}},"post":{"tags":["Admin Policy"],"summary":"Activate policy","description":"Activates a policy from back office.","operationId":"activatePolicy","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PolicyActivationRequest"}}},"required":true},"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyActivationResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyActivationResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyActivationResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyActivationResponse"}}}}}},"delete":{"tags":["Admin Policy"],"summary":"Deactivate policy","description":"Deactivates a policy from back office.","operationId":"deactivatePolicy","parameters":[{"name":"policyId","in":"query","description":"Policy identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":1003}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyDeactivationResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyDeactivationResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyDeactivationResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PolicyDeactivationResponse"}}}}}}},"/api/families/policies":{"get":{"tags":["User Policy"],"summary":"Get applied policies by family","description":"Returns policies currently applied to a family.","operationId":"getFamilyPolicies","parameters":[{"name":"familyId","in":"query","description":"Family identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":10}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}}}},"post":{"tags":["User Policy"],"summary":"Apply family policy","description":"Applies an activated policy to a specific family.","operationId":"applyFamilyPolicy","parameters":[{"name":"familyId","in":"query","description":"Family identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":10}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FamilyPolicyApplyRequest"}}},"required":true},"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}}}},"delete":{"tags":["User Policy"],"summary":"Remove family policy","description":"Removes an applied policy from a specific family.","operationId":"removeFamilyPolicy","parameters":[{"name":"familyId","in":"query","description":"Family identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":10},{"name":"policyId","in":"query","description":"Policy identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":1003}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FamilyPolicyChangeResponse"}}}}}}},"/api/lines/policies":{"patch":{"tags":["User Policy"],"summary":"Update line policies","description":"Bulk updates detailed policy settings for a specific line.","operationId":"updateLinePolicies","parameters":[{"name":"lineId","in":"query","description":"Line identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinePolicyUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LinePolicyUpdateResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LinePolicyUpdateResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LinePolicyUpdateResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LinePolicyUpdateResponse"}}}}}}},"/api/policies/all":{"get":{"tags":["Admin Policy"],"summary":"List all policies","description":"Returns all policies for admin including both active and inactive items.","operationId":"getAllPolicies","responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminPolicyResponse"}}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminPolicyResponse"}}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminPolicyResponse"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminPolicyResponse"}}}}}}}},"/api/lines/limits":{"get":{"tags":["User Policy"],"summary":"Get line limit policies","description":"Returns limit policy details for a specific line.","operationId":"getLimitPolicies","parameters":[{"name":"lineId","in":"query","description":"Line identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LimitPolicyResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LimitPolicyResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LimitPolicyResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LimitPolicyResponse"}}}}}}},"/api/lines/blocks":{"get":{"tags":["User Policy"],"summary":"Get line block policies","description":"Returns block policy details for a specific line.","operationId":"getBlockPolicies","parameters":[{"name":"lineId","in":"query","description":"Line identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlockPolicyResponse"}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlockPolicyResponse"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlockPolicyResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlockPolicyResponse"}}}}}}},"/api/lines/apps/usage":{"get":{"tags":["Admin Policy"],"summary":"Get app usage by line","description":"Returns app usage statistics for a specific line.","operationId":"getLineAppUsage","parameters":[{"name":"lineId","in":"query","description":"Line identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LineAppUsageResponse"}}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LineAppUsageResponse"}}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LineAppUsageResponse"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LineAppUsageResponse"}}}}}}}},"/api/lines/apps/policies":{"get":{"tags":["User Policy"],"summary":"Get app policies by line","description":"Returns app-level policy list for a specific line.","operationId":"getAppPolicies","parameters":[{"name":"lineId","in":"query","description":"Line identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppPolicyResponse"}}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppPolicyResponse"}}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppPolicyResponse"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppPolicyResponse"}}}}}}}},"/api/lines/applied-policies":{"get":{"tags":["User Policy"],"summary":"Get applied policies by line","description":"Returns currently applied policy list for a specific line.","operationId":"getAppliedPoliciesByLine","parameters":[{"name":"lineId","in":"query","description":"Line identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":101}],"responses":{"200":{"description":"Request succeeded","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}},"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AppliedPolicyResponse"}}}}}}}}},"components":{"schemas":{"PolicyActivationRequest":{"type":"object","description":"Request body for activating a policy","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier to activate","example":1003}},"required":["policyId"]},"PolicyActivationResponse":{"type":"object","description":"Result of policy activation","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier","example":1003},"active":{"type":"boolean","description":"Active flag after operation","example":true},"activatedAt":{"type":"string","description":"Activated timestamp","example":"2026-02-20T10:10:00"}}},"FamilyPolicyApplyRequest":{"type":"object","description":"Request body for applying a policy to a family","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier to apply","example":1002}},"required":["policyId"]},"FamilyPolicyChangeResponse":{"type":"object","description":"Result of family policy apply/remove operation","properties":{"familyId":{"type":"integer","format":"int64","description":"Family identifier","example":10},"policyId":{"type":"integer","format":"int64","description":"Policy identifier","example":1002},"status":{"type":"string","description":"Operation status","example":"APPLIED"},"processedAt":{"type":"string","description":"Processed timestamp","example":"2026-02-20T10:10:00"}}},"LinePolicyUpdateRequest":{"type":"object","description":"Request body for bulk line policy update","properties":{"blockPolicyEnabled":{"type":"boolean","description":"Whether line-level block policy is enabled","example":true},"dailyLimitMb":{"type":"integer","format":"int32","description":"Daily data limit in MB","example":1024,"minimum":0},"allowedAppPolicyIds":{"type":"array","description":"Allowed app policy identifiers","example":[11,12,13],"items":{"type":"integer","format":"int64"}}},"required":["allowedAppPolicyIds","blockPolicyEnabled","dailyLimitMb"]},"LinePolicyUpdateResponse":{"type":"object","description":"Result of line policy bulk update","properties":{"lineId":{"type":"integer","format":"int64","description":"Line identifier","example":101},"updatedPolicyCount":{"type":"integer","format":"int32","description":"Count of updated policy entries","example":5},"updatedAt":{"type":"string","description":"Update timestamp","example":"2026-02-19T12:00:00"}}},"ActivePolicyResponse":{"type":"object","description":"Active policy that can be applied to a family","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier","example":1001},"policyName":{"type":"string","description":"Policy name","example":"Night Usage Block"},"policyType":{"type":"string","description":"Policy type","example":"BLOCK"},"description":{"type":"string","description":"Policy description","example":"Blocks data usage from 22:00 to 06:00."}}},"AdminPolicyResponse":{"type":"object","description":"Policy item for admin policy list","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier","example":1001},"policyName":{"type":"string","description":"Policy name","example":"Night Usage Block"},"policyType":{"type":"string","description":"Policy type","example":"BLOCK"},"active":{"type":"boolean","description":"Whether this policy is active","example":true},"updatedAt":{"type":"string","description":"Last updated timestamp","example":"2026-02-20T10:10:00"}}},"LimitPolicyResponse":{"type":"object","description":"Limit policy detail for a specific line","properties":{"lineId":{"type":"integer","format":"int64","description":"Line identifier","example":101},"dailyLimitMb":{"type":"integer","format":"int32","description":"Daily data limit in MB","example":1024},"monthlyLimitMb":{"type":"integer","format":"int32","description":"Monthly data limit in MB","example":20480},"warningThresholdPercent":{"type":"integer","format":"int32","description":"Warning threshold percent","example":80}}},"BlockPolicyResponse":{"type":"object","description":"Block policy detail for a specific line","properties":{"lineId":{"type":"integer","format":"int64","description":"Line identifier","example":101},"blockAdultContent":{"type":"boolean","description":"Adult content block status","example":true},"blockRoaming":{"type":"boolean","description":"Roaming data block status","example":false},"blockPaidContent":{"type":"boolean","description":"Paid content block status","example":true}}},"LineAppUsageResponse":{"type":"object","description":"App usage statistics for a specific line","properties":{"lineId":{"type":"integer","format":"int64","description":"Line identifier","example":101},"appId":{"type":"integer","format":"int64","description":"App identifier","example":301},"appName":{"type":"string","description":"App name","example":"YouTube"},"usedMb":{"type":"integer","format":"int64","description":"Used data in MB","example":2450},"usagePercent":{"type":"integer","format":"int32","description":"Usage share percent","example":32}}},"AppPolicyResponse":{"type":"object","description":"App-level policy detail for a specific line","properties":{"appId":{"type":"integer","format":"int64","description":"App identifier","example":301},"appName":{"type":"string","description":"App name","example":"YouTube"},"policyType":{"type":"string","description":"Policy type","example":"LIMIT"},"enabled":{"type":"boolean","description":"Whether this app policy is enabled","example":true},"dailyLimitMb":{"type":"integer","format":"int32","description":"App daily limit in MB","example":500}}},"AppliedPolicyResponse":{"type":"object","description":"Applied policy detail","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier","example":1001},"policyName":{"type":"string","description":"Policy name","example":"Night Usage Block"},"policyType":{"type":"string","description":"Policy type","example":"BLOCK"},"appliedTarget":{"type":"string","description":"Applied target type","example":"LINE"},"targetId":{"type":"integer","format":"int64","description":"Target identifier (line or family)","example":101},"appliedAt":{"type":"string","description":"Applied timestamp","example":"2026-02-20T10:10:00"}}},"PolicyDeactivationResponse":{"type":"object","description":"Result of policy deactivation","properties":{"policyId":{"type":"integer","format":"int64","description":"Policy identifier","example":1003},"active":{"type":"boolean","description":"Active flag after operation","example":false},"deactivatedAt":{"type":"string","description":"Deactivated timestamp","example":"2026-02-20T10:10:00"}}}}}}